Add Dockerfile for CUDA 11.8.0 (#463)
* Add Dockerfile for CUDA 11.8.0

This is almost the same as the anaconda-pkg-build container, but derived from
Nvidia's CentOS 7 CUDA 11.8.0 image.

Higher versions of CUDA are supported by conda; see comment in the Dockerfile.

* Add in CUDA_VERSION option to container

* Move dockerfile location to support multiple archs later

* copy linux workflow file as a basis for linux-cuda workflow file

* Adapt existing Linux build action to handle CUDA container build

Remove concourse stuff as concourse is no more

Adapt paths, image names as required

Leave in multi-platform functionality to be able to add more
architectures later if required

* Correct the CUDA version, needs patch version too

* Ignore hadolint warning

The hadolint error message was, "SC3045 warning: In POSIX sh, ulimit -n is
undefined.". ulimit -n 1024 both works and is necessary for the base container
(tested with nvidia/cuda:11.8.0-devel-centos7).

* remove -cuda from the repo name

These repos don't exist. We can differentiate from the existing one with a suffix tag.

Co-authored-by: Daniel Bast <2790401+dbast@users.noreply.github.com>

* Add cuda suffix

To differentiate from the existing non-cuda container.

Co-authored-by: Daniel Bast <2790401+dbast@users.noreply.github.com>

---------

Co-authored-by: Daniel Bast <2790401+dbast@users.noreply.github.com>